create reducer
