+++ src/hooks.tsx
@@ -1,6 +1,9 @@
+import { useEffect } from 'react';
+
 import { useOidcJwtContext } from './OidcJwtProvider';
 import { ClaimsBase, Params } from './store';
-import { State, usePromiseResult } from './utils/usePromiseResult';
+import { State, useAuthReducer } from './utils/useAuthReducer';
+import { usePromiseResult } from './utils/usePromiseResult';
 
 interface OidcAuthControls {
   authorize(params?: Params): void;
@@ -24,26 +27,40 @@ function useAuthUserInfo<T>(): State<T | null> {
   const { useStore } = useOidcJwtContext();
   const getUserInfo = useStore(state => state.methods.getUserInfo);
   const { hasValidSession } = useAuthSessionInfo();
+  const [state, dispatch] = useAuthReducer<T>();
 
-  return usePromiseResult<T | null>(() => {
+  const value = usePromiseResult<T | null>(() => {
     if (!hasValidSession) {
       return Promise.resolve(null);
     }
     return getUserInfo<T>();
   }, []);
+
+  useEffect(() => {
+    dispatch({ type: 'SET_DATA', payload: value });
+  }, [dispatch, value]);
+
+  return state;
 }
 
 function useAuthAccessClaims<T extends ClaimsBase>(): State<T | null> {
   const { useStore } = useOidcJwtContext();
   const getAccessToken = useStore(state => state.methods.getAccessToken);
   const { hasValidSession } = useAuthSessionInfo();
+  const [state, dispatch] = useAuthReducer<T>();
 
-  return usePromiseResult<T | null>(() => {
+  const value = usePromiseResult<T | null>(() => {
     if (!hasValidSession) {
       return Promise.resolve(null);
     }
     return getAccessToken<T>().then(info => info?.claims ?? null);
   }, []);
+
+  useEffect(() => {
+    dispatch({ type: 'SET_DATA', payload: value });
+  }, [dispatch, value]);
+
+  return state;
 }
 
 function useAuthSessionInfo(): OidcAuthSessionInfo {
